requirements.txt
Flask
bcrypt
regex
mysql-connector-python
colorblind
emoji
opencv-python
Pillow
numpy
deepface
python-facepp